requirements.txt
streamlit==0.72.0
numpy==1.19.4
pandas==0.24.1
Pillow==8.0.1